requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.6.2
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.3
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dj-config-url==0.1.1
dj-database-url==2.2.0
Django==3.2.25
django-environ==0.11.2
django-plotly-dash==2.2.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dpd-components==0.1.0
dpd-static-support==0.0.5
Faker==25.5.0
flake8==7.0.0
Flask==3.0.3
gunicorn==20.1.0
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
plotly==5.22.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
retrying==1.3.4
six==1.16.0
sqlparse==0.5.0
tenacity==8.3.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.1
Werkzeug==3.0.3
whitenoise==6.6.0
zipp==3.19.2
